+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Chg #81: Make category parameter in `TranslatorInterface::addCategorySources()` variadic, and remove 
  `TranslatorInterface::addCategorySource()` method (@vjik)
 - New #82: Add parameter `$defaultCategory` to `Translator` constructor (@vjik)
+- Chg #84: In `TranslatorInterface` rename method `withCategory()` to `withDefaultCategory()` (@vjik)
 
 ## 1.1.1 September 09, 2022