+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - New #82: Add parameter `$defaultCategory` to `Translator` constructor (@vjik)
 - Chg #87: Fix package configuration, remove default category source, change default locale to `en_US` (@vjik) 
 - Chg #84: In `TranslatorInterface` rename method `withCategory()` to `withDefaultCategory()` (@vjik)
+- Chg #90: Simplified category sources config (@rustamwin)
 
 ## 1.1.1 September 09, 2022